requirements.txt
faiss-cpu
numpy
pytest
fastapi
uvicorn